feat: move create-module logic to brickyard
