+++ lib/brickyard/index.js
@@ -6,6 +6,7 @@ const semver = require('semver')
 const Bluebird = require('bluebird')
 const glob = require('glob')
 const fs = Bluebird.promisifyAll(require('fs'))
+const mkdirp = Bluebird.promisify(require('mkdirp'))
 const gulp = require('gulp')
 const del = require('del')
 
@@ -38,10 +39,23 @@ function getPath(file) {
 	return path.isAbsolute(file) ? file : path.join(process.cwd(), file)
 }
 
+function getPackageJson(name, type) {
+	const rs = {}
+	rs.name = name
+	rs['brickyard-module-type'] = type
+	rs.version = '0.0.1'
+	rs.description = `${type} for ${name}`
+	rs.main = 'index.js'
+	rs.author = 'brickyard-developer'
+	rs.license = 'ISC'
+	return rs
+}
+
 class Brickyard extends events.EventEmitter {
 
 	constructor() {
 		super()
+		this.Class = Brickyard
 		this.allModules = new BrickyardArray()
 		this.modules = new BrickyardArray()
 
@@ -244,6 +258,23 @@ class Brickyard extends events.EventEmitter {
 		}
 	}
 	/** Exports Start*/
+	static async createModule(type, dir, name) {
+		const candidates = ['frontend', 'backend', 'buildtask', 'plan']
+		if (candidates.indexOf(type) === -1) {
+			throw new Error(`module type must be ${candidates}`)
+		}
+		const mdName = name || path.basename(dir)
+		if (!/[\w-\d]*/.test(mdName)) {
+			throw new Error(`${name} is not a valid module name`)
+		}
+		await mkdirp(dir)
+		const conf = getPackageJson(mdName, type)
+		const configContent = `${JSON.stringify(conf, null, 2)}\n`
+		const configPath = path.join(dir, 'package.json')
+		await fs.writeFileAsync(configPath, configContent)
+		await fs.writeFileAsync(path.join(dir, 'index.js'), '')
+		return new BrickyardModules(configPath)
+	}
 
 	/*
 	 * ensure brickyard-cli version is greater than the param
