+++ lib/gulpfile.js
@@ -1,6 +1,3 @@
-const fs = require('fs')
-const path = require('path')
-const mkdirp = require('mkdirp')
 const childProcess = require('child_process')
 const Bluebird = require('bluebird')
 const _ = require('lodash')
@@ -155,33 +152,9 @@ gulp.task('run', (() => async () => {
 	}
 })())
 
-function getPackageJson(name, type) {
-	const rs = {}
-	rs.name = name
-	rs['brickyard-module-type'] = type
-	rs.version = '0.0.1'
-	rs.description = `${type} for ${name}`
-	rs.main = 'index.js'
-	rs.author = 'brickyard-developer'
-	rs.license = 'ISC'
-	return rs
-}
-
 gulp.task('create-module', (() => async () => {
-	const { type, dir } = argv
+	const { type, dir, name } = argv
 
-	const candidates = ['frontend', 'backend', 'buildtask', 'plan']
-	if (candidates.indexOf(type) === -1) {
-		throw new Error(`module type must be ${candidates}`)
-	}
-	const name = argv.name || path.basename(argv.dir)
-	if (!/[\w-\d]*/.test(name)) {
-		throw new Error(`${name} is not a valid module name`)
-	}
-	mkdirp.sync(dir)
-	const conf = getPackageJson(name, type)
-	const confContent = `${JSON.stringify(conf, null, 2)}\n`
-	fs.writeFileSync(path.join(dir, 'package.json'), confContent)
-	fs.writeFileSync(path.join(dir, 'index.js'), '')
-	console.log(`${type} module ${name} created at ${dir}`)
+	const md = await brickyard.Class.createModule(type, dir, name)
+	console.log(`${md.type} module ${md.name} created at ${dir}`)
 })())
