+++ lib/harmony.js
@@ -1,5 +1,5 @@
 /*
- * 由于brickyard-cli用了async function
+ * 由于brickyard-cli用了async function，需要用node@^7.6.0或node7的harmony模式
  * 本模块用于使当前程序运行于harmony模式，注意事项
  * 1. 本模块通过child_process.spawn()方法实现
  * 2. 引入本模块前的模块，和引入本模块的当前模块，不能使用harmony语法
